streamlit_app/streamlit/requirements.txt
streamlit
pandas
snowflake-sqlalchemy
plotly
python-dotenv
streamlit-autorefresh